% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipeline_accessions.R
\name{accessions_to_use}
\alias{accessions_to_use}
\title{Search / define a list of accessions}
\usage{
accessions_to_use(
  user_specified = character(),
  organisms = "Homo sapiens",
  massive_search = TRUE,
  libtypes = "Ribosome profiling",
  rpfdb = "Ribosome profiling" \%in\% libtypes
)
}
\arguments{
\item{user_specified}{character, default character(), user specified accessions}

\item{organisms}{character vector, default "all" (Use all organisms found).
Else binomial latin name with capital letter for genus: "Homo sapiens" etc.}

\item{massive_search}{logical, default TRUE. Massive accession search over
NCBI using search terms defined over organisms and libtypes.}

\item{libtypes}{character, default "Ribosome profiling",
which library types to use: Other examples could be
"Guide seq", "mRNA seq", "CHIP seq", etc.}

\item{rpfdb}{logical, default "Ribosome profiling" \%in\% libtypes.
If so will get all rpfdb accessions for given organisms too.}
}
\description{
There are the primary ways:\cr
\itemize{
\item Define the accessions yourself
\item massive search over NCBI
\item Get from rpfdb
}
}
