% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipeline_metadata.R
\name{add_new_data}
\alias{add_new_data}
\title{Add new data to the temp_metadata}
\usage{
add_new_data(
  accessions,
  config,
  organisms = "all",
  google_url = config$google_url,
  complete_metadata = config$complete_metadata,
  LibraryStrategy = c("RNA-Seq", "miRNA-Seq", "OTHER"),
  LibraryLayouts = c("SINGLE", "PAIRED"),
  Platforms = "ILLUMINA",
  open_editor = interactive(),
  open_google_sheet = interactive(),
  organism_name_cleanup = TRUE
)
}
\description{
Add new data to the temp_metadata
}
