% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipeline_flags.R
\name{add_step_to_pipeline}
\alias{add_step_to_pipeline}
\title{Add additional step to pipeline}
\usage{
add_step_to_pipeline(
  config,
  short_name,
  FUN,
  group_name = name_of_function(FUN),
  tail_or_head = "tail"
)
}
\arguments{
\item{config}{list, a NGS pipeline object}

\item{short_name}{character, the flag id, a short identifier (differential expression could be: difexp etc)}

\item{FUN}{a function, the to be added}

\item{group_name}{character, name of the function FUN.}

\item{tail_or_head}{character, c("tail", "head")\link{1}. Default is to append last, can also front append.}
}
\value{
list, the updated config object
}
\description{
Add additional step to pipeline
}
