% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/translon_id.R
\name{as_translon_info_list}
\alias{as_translon_info_list}
\title{Convert GRangesList of translons to generic strings}
\usage{
as_translon_info_list(grl, assembly = unique(genome(grl)))
}
\arguments{
\item{grl}{a GRangesList}

\item{assembly}{character, the assembly name (e.g. GRCh38)}
}
\value{
a character vector
}
\description{
This is input for make_translon_id
}
\examples{
grl <- makeGRangesListFromCharacter("chr22:101231230-101231500:+;chr22:101240000-101240962:+")
genome(grl) <- "GRCh38"
translon_string <- as_translon_info_list(grl)
translon_string
}
