% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fastq_helpers.R
\name{barcode_detector_pipeline}
\alias{barcode_detector_pipeline}
\title{NGS barcode detector
Detects both 5' and 3' barcodes}
\usage{
barcode_detector_pipeline(pipeline, redownload_raw_if_needed = TRUE)
}
\arguments{
\item{pipeline}{a pipeline object}

\item{redownload_raw_if_needed}{logical, default TRUE.
For additional stats redownload fastq if removed.}
}
\value{
data.table with stats for all runs in experiment
}
\description{
NGS barcode detector
Detects both 5' and 3' barcodes
}
\examples{
config <- pipeline_config()
pipelines <- pipeline_init_all(config, only_complete_genomes = TRUE, gene_symbols = FALSE, simple_progress_report = FALSE)
pipeline <- pipelines[["PRJNA634994"]]
# barcode_detector_pipeline(pipeline, FALSE)
# debug(detect_adapter_and_trim)
# pipeline$study <- pipeline$study[Run == "SRR10846527"]
barcode_detector_pipeline(pipeline, TRUE)
}
