% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipeline_config.R
\name{config_substeps}
\alias{config_substeps}
\title{Configure which flags are done in each function}
\usage{
config_substeps(
  flags,
  mode,
  preset = "Ribo-seq",
  substeps = preset_substeps(preset, flags, mode)
)
}
\arguments{
\item{flags}{character vector, names of flags}

\item{mode}{= c("online", "local")\link{1}. "online" will assume project IDs for
online repository (SRA, ENA, PRJ etc). Local means local folders as accessions.}

\item{substeps}{a list of character vectors, where each sub list
is the steps that happen in that sub component of the pipeline}
}
\value{
a list, returns the validated 'substeps' argument.
}
\description{
Configure which flags are done in each function
}
