% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipeline_metadata.R
\name{curate_metadata}
\alias{curate_metadata}
\title{Curate metadata for massive_NGS_pipe}
\usage{
curate_metadata(
  accessions,
  config,
  organisms = "all",
  google_url = config$google_url,
  complete_metadata = config$complete_metadata,
  LibraryLayouts = c("SINGLE", "PAIRED"),
  Platforms = "ILLUMINA",
  step_mode = FALSE,
  open_google_sheet = interactive(),
  fix_loop = TRUE
)
}
\arguments{
\item{accessions}{character vector, all candidate accession numbers, allowed types:\cr
\itemize{
\item Biorpoject ID (Only numbers)\cr
\item Bioproject accession (PRJ)\cr
\item SRA study (SRA)\cr
\item ENA study (ERA)\cr
\item GEO study (GSE)\cr
\item NULL (not "NULL", but NULL), don't add new accession, only look if you
current google sheet or local csv is ready.
}}

\item{config}{list, output from pipeline_config(), the global config for your
NGS pipeline}

\item{organisms}{character vector, default "all" (Use all organisms found).
Else binomial latin name with capital letter for genus: "Homo sapiens" etc.}

\item{google_url}{character, default config$google_url. If not NULL,
will use the google sheet to check for updated metadata.}

\item{complete_metadata}{path, default file.path(project_dir, "RFP_FINAL_LIST.csv").
The list of final candidates that are checked and have unique rows per bioproject}

\item{LibraryLayouts}{character vector, default c("SINGLE", "PAIRED"),
either or both of: c("SINGLE", "PAIRED")}

\item{Platforms}{character vector, default: "ILLUMINA". The sequencer technologies allowed.}

\item{step_mode}{logical, default FALSE. If TRUE, actives browser() and lets you go through
each step in debug mode for full control.}

\item{fix_loop}{logical, default TRUE. If TRUE, will create a loop in
the end that runs until your metadata is valid. If you
only want to updata csv / google docs with new data and not try to complete
now, set this to FALSE.}
}
\value{
logical, TRUE if you had unique rows per accession per organism
}
\description{
Runs an interactive while loop until your metadata is curated and have this property:
Per accession, your study must have a unique set of rows per organism.
Such that ORFik experiment can understand what is what:
(what is Wild type vs experimental), which are replicates etc.
The uniqueness is defined from column LIBRARYTYPE to column INHIBITOR
}
