% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/google_API.R
\name{default_sheets}
\alias{default_sheets}
\title{Get pre-existing or create a new google sheet for metadata}
\usage{
default_sheets(project_dir, id = 1, sheet_name = basename(project_dir))
}
\arguments{
\item{project_dir}{where will specific pipeline outputs be put}

\item{id}{numeric, the row index of id. Default is 1.}

\item{sheet_name}{character, default basename(project_dir)}
}
\value{
a string of full url.
}
\description{
Get pre-existing or create a new google sheet for metadata
}
