% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/demultiplex.R
\name{demultiplex}
\alias{demultiplex}
\title{Given ..., demultiplex to fastq}
\usage{
demultiplex(path, csv, demultiplex_tool_bin = "~/bin/demultiplexer")
}
\arguments{
\item{path}{path to directory of}

\item{csv}{path to csv file of}
}
\value{
invisible(NULL), files saved to disc
}
\description{
Given ..., demultiplex to fastq
}
