% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bam_utils.R
\name{detect_strand_mode}
\alias{detect_strand_mode}
\title{Detect if R1 or R2 of read pair is primary read direction}
\usage{
detect_strand_mode(
  R1,
  R2,
  genomeDir = "~/livemount/Bio_data/references/homo_sapiens/STAR_index/genomeDir/",
  nreads = 1e+06,
  tx = loadRegion(list.files(dirname(dirname(genomeDir)), ".db$", full.names = TRUE)[1],
    "tx"),
  td = tempdir(),
  threads = 20,
  star = STAR.install(),
  keepGenomeLoaded = c("LoadAndRemove", "LoadAndKeep", "NoSharedMemory")[1]
)
}
\arguments{
\item{R1}{path to R1 fasta/fastq file}

\item{R2}{path to R2 fasta/fastq file}

\item{genomeDir}{path to STAR index of genome, default:
"~/livemount/Bio_data/references/homo_sapiens/STAR_index/genomeDir/"}

\item{nreads}{numeric, default 1e6 (number of reads to use)}

\item{tx}{GRangesList, the transcripts to count overlaps on}

\item{td}{path to tempdir to use, default tempdir()}

\item{threads}{numeric, default 20}

\item{star}{path to STAR, default STAR.install()}

\item{keepGenomeLoaded}{character, default c("LoadAndRemove", "LoadAndKeep", "NoSharedMemory")\link{1}}
}
\value{
a named numeric (1 or 2), 1 means R1 is primary. Name gives ratio of overlaps of R1 / R2.
So 0.07 means R2 is bigger and 7\% overlaps in R1 compared to R2.
}
\description{
Detect if R1 or R2 of read pair is primary read direction
}
\examples{
genomeDir <- "~/livemount/Bio_data/references/mus_musculus/STAR_index/genomeDir/"
files <- c("~/livemount/Bio_data/raw_data/RNA-seq/PRJNA985729-mus_musculus/SRR24972841_1.fastq",
 "~/livemount/Bio_data/raw_data/RNA-seq/PRJNA985729-mus_musculus/SRR24972841_2.fastq")
R1 <- files[1]
R2 <- files[2]
#detect_strand_mode(R1, R2, genomeDir)
}
