% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipeline_utils.R
\name{dirExp}
\alias{dirExp}
\title{Get path to dir type of exp}
\usage{
dirExp(exp, rel_dir = "aligned")
}
\arguments{
\item{exp}{ORFik experiment or character name of experiment,
a ORFik experiment name of folder to go to, e.g. "GSE91068-saccharomyces_cerevisiae".}

\item{rel_dir}{character, relative directory, "aligned", "trim", c("", "processed") or "raw"}
}
\value{
a character path
}
\description{
Navigate to a directory in the Files pane.
The contents of that directory will be listed and shown in the Files pane.
}
