% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipeline_validate.R
\name{docker_copy_done_experiments}
\alias{docker_copy_done_experiments}
\title{For docker usage copy}
\usage{
docker_copy_done_experiments(
  config,
  pipelines = pipeline_init_all(config),
  new_exp_dir = "~/livemount/Bio_data/ORFik_experiments/",
  docker_conversion = "'s/livemount\\\\///g'",
  merged_by_organism = TRUE
)
}
\arguments{
\item{config}{a config object}

\item{pipelines}{pipeline objects, default pipeline_init_all(config)}

\item{new_exp_dir}{character, default "~/livemount/Bio_data/ORFik_experiments/"}

\item{docker_conversion}{the sed conversion string: 's/livemount\///g'}
}
\value{
logical, TRUE if sucessful for all.
}
\description{
If you have download docker and release docker.
This is useful to copy files from download to release
}
