% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/google_API.R
\name{download_fastq_google_drive}
\alias{download_fastq_google_drive}
\title{Get all fastq files from google drive}
\usage{
download_fastq_google_drive(
  drive_url = "https://drive.google.com/drive/folders/1N_Tyig3-IxMIe8Im24wBSNiS5vSohi91",
  format_pattern = "fastq\\\\.gz$",
  output_dir = "~/livemount/Bio_data/raw_data/autophagy_timecourse2/",
  email = "hakontjeldnes@gmail.com",
  overwrite = FALSE
)
}
\arguments{
\item{drive_url}{character, url to drive folder (not individual files), example:
"https://drive.google.com/drive/folders/1N_Tyig3-IxMIe8Im24wBSNiS5vSohi91"}

\item{output_dir}{character, output directory, will be create if not existing}

\item{email}{character, email of user, default: \email{hakontjeldnes@gmail.com}}

\item{overwrite}{logical, default FALSE. If FALSE, delete existing files,
if TRUE dont download existing files.}

\item{pattern}{character pattern of files, default "fastq\\.gz$"}
}
\value{
invisible(NULL)
}
\description{
Get all fastq files from google drive
}
