% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/demultiplex.R
\name{excel_Wiktoria_to_csv}
\alias{excel_Wiktoria_to_csv}
\title{Convert Wiktoria template excel sheet to csv controlfile}
\usage{
excel_Wiktoria_to_csv(
  path =
    "~/livemount/Bio_data/raw_data/vaccine_mod_bnt_1225-homo_sapiens_x_sars_cov2/multiplexed/Ribosome-profiling_251010.xlsx",
  adapter = NULL
)
}
\arguments{
\item{path}{path to xlsx file}

\item{adapter}{NULL, else character, if specified add adapters to avoid needing to check it.}
}
\value{
a data.table, also saved to disc at file.path(dirname(path), "controlfile.csv")
}
\description{
Convert Wiktoria template excel sheet to csv controlfile
}
\examples{
path <- "~/livemount/Bio_data/raw_data/vaccine_mod_bnt_1225-homo_sapiens_x_sars_cov2/multiplexed/Ribosome-profiling_251010.xlsx"
adapter <- c(Illumina_TRUESEQ = "AGATCGGAAGAG")
excel_Wiktoria_to_csv(path, adapter)
}
