% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/google_API.R
\name{export_discord_cache_file}
\alias{export_discord_cache_file}
\title{Given a webhook, create cache file}
\usage{
export_discord_cache_file(filepath = "~/.cache/discordr/r_to_discord_config")
}
\arguments{
\item{filepath}{character path, default "~/.cache/discordr/r_to_discord_config"}
}
\value{
Returns the path you specified as input if sucessful.
}
\description{
Will append if webhooks exist in file already.
}
