% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipeline_utils.R
\name{file_statistics_all_types}
\alias{file_statistics_all_types}
\title{Get massiveNGSpipe related file counts and total drive usage}
\usage{
file_statistics_all_types(config = ORFik::config(), include_references = TRUE)
}
\arguments{
\item{config}{config = ORFik::config()}

\item{include_references}{= TRUE}
}
\value{
a data.table
}
\description{
Get massiveNGSpipe related file counts and total drive usage
}
\examples{
config <- ORFik::config()
config_RNA <- file.path(config, "RNA-seq")
config_SSU <- file.path(config, "SSU")
config_disome <- file.path(config, "disome")
names(config_RNA) <- names(config_SSU) <- names(config_disome) <- names(config)
info_dt <- rbindlist(
list(RFP =    file_statistics_all_types(config),
     RNA =    file_statistics_all_types(config_RNA, FALSE),
     SSU =    file_statistics_all_types(config_SSU, FALSE),
     disome = file_statistics_all_types(config_disome, FALSE)), idcol = TRUE)
 info_dt[, sum(total_size_GB), by = .id]
}
