% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipeline_config.R
\name{flag_grouping}
\alias{flag_grouping}
\title{Define which flags are done in each function}
\usage{
flag_grouping(flags, substeps = preset_grouping(flags))
}
\arguments{
\item{flags}{character vector, of paths with names as flag steps and
attribute called "grouping" with length same as flag with grouping information,
where group names are the group functions.}

\item{substeps}{a list of character vectors, where each sub list
is the steps that happen in that sub component of the pipeline}
}
\value{
a list, returns the validated 'substeps' argument.
}
\description{
Define which flags are done in each function
}
