% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipeline_checklist.R
\name{format_checklist}
\alias{format_checklist}
\title{Format a checklist data.table as a compact printable table}
\usage{
format_checklist(tab)
}
\arguments{
\item{tab}{data.table, as returned by pipeline_checklist(print = FALSE)}
}
\value{
character, one formatted line per stage
}
\description{
Format a checklist data.table as a compact printable table
}
