% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fst_utils.R
\name{generateAllTranscriptsMatrix}
\alias{generateAllTranscriptsMatrix}
\title{General fst all samples tables}
\usage{
generateAllTranscriptsMatrix(
  df,
  trxs = names(grl),
  output_dir = RiboCrypt::collection_dir_from_exp(df),
  grl = loadRegion(df, "mrna"),
  libnames = bamVarName(df),
  verbose = TRUE,
  BPPARAM = BiocParallel::bpparam()
)
}
\arguments{
\item{df}{ORFik experiment of all samples per species}

\item{trxs}{transcript ids, default names(grl)}

\item{output_dir}{RiboCrypt::collection_dir_from_exp(df)}

\item{grl}{loadRegion(df, "transcript")}

\item{libnames}{= bamVarName(df), the names of columns for fst}

\item{verbose}{logical, TRUE. Print name of current tx.}

\item{BPPARAM}{BiocParallel::bpparam(), for single thread,
do: BiocParallel::SerialParam()}
}
\value{
invisible(NULL), results saved to disc.
}
\description{
General fst all samples tables
}
