% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bam_utils.R
\name{get_expanded_alignment_metrics}
\alias{get_expanded_alignment_metrics}
\title{Recreate alignment report for the decollapsed reads}
\usage{
get_expanded_alignment_metrics(fasta_file, bam_file)
}
\description{
Recreate alignment report for the decollapsed reads
}
