% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bam_utils.R
\name{get_expanded_alignment_metrics_exp}
\alias{get_expanded_alignment_metrics_exp}
\title{Recreate alignment report for the decollapsed reads for whole exp}
\usage{
get_expanded_alignment_metrics_exp(
  df,
  fasta_dir = file.path(dirname(libFolder(df)), "trim", "SINGLE"),
  BPPARAM = MulticoreParam(8, exportglobals = FALSE)
)
}
\arguments{
\item{df}{ORFik experiment}

\item{fasta_dir}{path of collapsed files}

\item{BPPARAM}{a BPPARAM object}
}
\value{
a data.table, with uncollapsed alignment statistics
}
\description{
Recreate alignment report for the decollapsed reads for whole exp
}
