% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/google_API.R
\name{google_drive_list_files}
\alias{google_drive_list_files}
\title{List google files on drive wrapper}
\usage{
google_drive_list_files(
  drive_url,
  email = gargle_mail_safe(),
  format_pattern = "fastq\\\\.gz$",
  recursive = TRUE
)
}
\arguments{
\item{drive_url}{character, url to drive folder (not individual files), example:
"https://drive.google.com/drive/folders/1N_Tyig3-IxMIe8Im24wBSNiS5vSohi91"}

\item{email}{character, email of user, default: \email{hakontjeldnes@gmail.com}}

\item{recursive}{logical, default TRUE. Search sub folders for format files.}
}
\value{
a dribble of googl drive information
}
\description{
List google files on drive wrapper
}
