% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_sra.R
\name{install_aws}
\alias{install_aws}
\title{Install aws (AWS CLI v2), if not already installed}
\usage{
install_aws(path = "bin/aws")
}
\arguments{
\item{path}{path to the aws binary, relative to the home directory,
default "bin/aws".}
}
\value{
path to runnable aws, only ever returned if it exists.
}
\description{
Install aws (AWS CLI v2), if not already installed
}
