% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipeline_logs.R
\name{is_progressing}
\alias{is_progressing}
\title{Check if anything happened since given time}
\usage{
is_progressing(config, max_time_since_update = 1)
}
\arguments{
\item{config}{a pipeline_config object}

\item{max_time_since_update}{numeric, in days, default 1.
1 Hour is 0.041}
}
\value{
logical, FALSE if nothing happend since max time
}
\description{
Check if anything happened since given time
}
