% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipeline_logs.R
\name{last_session_errors}
\alias{last_session_errors}
\title{Get last active session error}
\usage{
last_session_errors(config)
}
\arguments{
\item{config}{a list from a massiveNGSpipe::pipeline_config call}
}
\value{
a character vector of errors
}
\description{
Get last active session error
}
