% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipeline_progress.R
\name{last_session_stil_active}
\alias{last_session_stil_active}
\title{Is the last call to run_pipeline still running / zombie process}
\usage{
last_session_stil_active(config)
}
\arguments{
\item{config}{a pipeline config}
}
\value{
logical, TRUE / FALSE
}
\description{
Is the last call to run_pipeline still running / zombie process
}
