% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipeline_logs.R
\name{last_update_diff}
\alias{last_update_diff}
\title{Get last pipeline update, in days relative to now}
\usage{
last_update_diff(config, units = "days")
}
\arguments{
\item{units}{character, default "days"}

\item{a}{config object}

\item{flagdir}{character, default: dirname(config$flag\link{1}),
path to flag directory}
}
\value{
character, a date string
}
\description{
Get last pipeline update, in days relative to now
}
