% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipeline_logs.R
\name{last_update_message}
\alias{last_update_message}
\title{Get last update time message}
\usage{
last_update_message(config, units = "hours")
}
\arguments{
\item{config}{a list, the pipeline config}

\item{units}{= "hours"}
}
\value{
invisible(NULL), only cat prints the message
}
\description{
Get last update time message
}
