% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipeline_flags.R
\name{libtype_flags}
\alias{libtype_flags}
\title{Define the library specific steps}
\usage{
libtype_flags(preset, mode = "online")
}
\arguments{
\item{preset}{character, default "Ribo-seq".
Alternatives: c("Ribo-seq", "RNA-seq", "disome", "empty")}

\item{mode}{either of c("online", "local")}
}
\value{
a character vector with names being the grouping in functions
}
\description{
This is called flags
}
