% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipeline_preset_steps.R
\name{libtype_long_to_short}
\alias{libtype_long_to_short}
\title{Convert library type from long name to short name}
\usage{
libtype_long_to_short(long)
}
\arguments{
\item{long}{a character of long names}
}
\value{
a character of short names
}
\description{
Convert library type from long name to short name
}
