% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipeline_preset_steps.R
\name{libtype_short_to_long}
\alias{libtype_short_to_long}
\title{Convert library type from short name to long name}
\usage{
libtype_short_to_long(short)
}
\arguments{
\item{short}{a character of short names}
}
\value{
a character of long names
}
\description{
Convert library type from short name to long name
}
