% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipeline_utils.R
\name{list.experiments.list}
\alias{list.experiments.list}
\title{Get all experiment names to search with auto completion}
\usage{
list.experiments.list(
  pattern = "*",
  libtypeExclusive = NULL,
  dir = ORFik::config()["exp"]
)
}
\arguments{
\item{pattern}{allowed patterns in experiment file name:
default ("*", all experiments)}

\item{libtypeExclusive}{search for experiments with exclusivly this
libtype, default (NULL, all)}

\item{dir}{directory for ORFik experiments: default:
ORFik::config()["exp"], which by default is:
"~/Bio_data/ORFik_experiments/"}
}
\value{
a named list of characters,
so you can search experiment names with auto completion
}
\description{
Get all experiment names to search with auto completion
}
