% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipeline_class.R
\name{list_files_of_type}
\alias{list_files_of_type}
\title{Get files of study in dir type}
\usage{
list_files_of_type(pipeline, type = "fastq", full_names = FALSE)
}
\description{
Get files of study in dir type
}
