% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipeline_class.R
\name{list_files_of_type_missing}
\alias{list_files_of_type_missing}
\title{Get missing files of study in dir type}
\usage{
list_files_of_type_missing(pipelines, type = "fastq")
}
\description{
Get missing files of study in dir type
}
