% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipeline_preset_steps.R
\name{make_additional_formats}
\alias{make_additional_formats}
\title{Make additional formats for an experiment}
\usage{
make_additional_formats(df_ref, exp_name, out_dir_exp)
}
\arguments{
\item{df_ref}{an ORFik experiment with only 1 libtype, giving info for new.}

\item{exp_name}{name of experiment to be made}

\item{out_dir_exp}{folder path, where study files are found.}
}
\value{
invisible(NULL)
}
\description{
Creates a new experiment and new file formats.
This is used for custom merge experiments mostly like all samples
per organism.
}
