% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipeline_preset_steps.R
\name{make_additional_formats_internal}
\alias{make_additional_formats_internal}
\title{Make additional formats for an experiment}
\usage{
make_additional_formats_internal(
  df,
  libtype_df_ref = identical(libraryTypes(df), "RFP")
)
}
\arguments{
\item{df_ref}{an ORFik experiment}
}
\value{
invisible(NULL)
}
\description{
With pre-created experiment
}
