% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipeline_merge.R
\name{make_gtf_from_sequences}
\alias{make_gtf_from_sequences}
\title{Make gtf from fasta sequence files}
\usage{
make_gtf_from_sequences(fasta_files, type = "cds_only", strand = "+")
}
\arguments{
\item{fasta_files}{character, paths to fasta files}

\item{type}{character, default "cds_only", how ranges are defined TODO: implement
IRangesList input for non 1 cds start.}

\item{strand}{character, default "+", ranges are + stranded, else user defined of
length 1 or length of fasta files.}
}
\value{
paths to gtfs, or NULL if length of fasta_files is 0.
}
\description{
For each fasta, make a gtf, by default type is "cds_only", i.e. whole sequence is
cds.
}
