% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipeline_merge.R
\name{make_mixed_genome}
\alias{make_mixed_genome}
\title{Merge genome for mixed genome to align against}
\usage{
make_mixed_genome(
  species,
  custom_sequences,
  custom_gtfs = NULL,
  primary_organism = stringr::str_to_title(gsub("_", " ", species[1])),
  ref_dir = ORFik::config()["ref"],
  contaminants_sequences = "primary_only",
  gene_symbols = "primary_only"
)
}
\arguments{
\item{species}{character pre-existing genomes in the ref_dir: use basename (e.g. homo_sapiens)}

\item{custom_sequences}{character, path to custom fasta sequences}

\item{custom_gtfs}{character, default NULL, path to custom fasta sequences if existing}

\item{primary_organism}{character, scientific name (e.g. "Homo sapiens").
The name of the primary organism, so that biomart can search for symbols etc.}

\item{ref_dir}{location of existing references, default: ORFik::config()\link{"ref"}}

\item{contaminants_sequences}{path to STAR index to copy for contaminants, default "primary_only". Copies
the contaminants for species\link{1}}

\item{gene_symbols}{path to gene symbol fst to copy, default "primary_only". Copies
the fst for species\link{1}}
}
\value{
invisible(NULL), results saved to disc
}
\description{
Merge genome for mixed genome to align against
}
