% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/translon_id.R
\name{make_translon_id}
\alias{make_translon_id}
\title{Make translon id}
\usage{
make_translon_id(translon_info_list, short_len = 10)
}
\arguments{
\item{translon_info_list}{a list translon information needed of length 5.}

\item{short_len}{integer, default 10}
}
\value{
character vector of length equal to \code{lengths(translon_info_list)[2]}
}
\description{
Make the readable, stable ID:
\code{TRN:asm:ctg:Lk:strand:short-hash}
where 'short-hash' = first 'short_len' chars of
\code{sha512t24u(canonical_string$region_string)}, i.e string of
string of genomic coordinates without chromosome or strand.
}
\examples{
grl <- makeGRangesListFromCharacter("chr22:101231230-101231500:+;chr22:101240000-101240962:+")
genome(grl) <- "GRCh38"
as_translon_info_list <- as_translon_info_list(grl)
as_translon_info_list
make_translon_id(as_translon_info_list)
# output: "TRN:GRCh38:chr22:L3:+:AB3FZK2HQY"   # (your hash will differ)
make_translon_machine_id(canon)
# output: "TRN:V1:sha512t24u:Oqk...<longer url-safe digest>"
}
