% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/massiveNGSpipe-package.R
\docType{package}
\name{massiveNGSpipe-package}
\alias{massiveNGSpipe}
\alias{massiveNGSpipe-package}
\title{massiveNGSpipe: Pipeline for automatic NGS data download and analysis over a massive set of studies}
\description{
R package for automatic NGS data download and analysis over a massive set of studies. With main use cases: Find studies you want, download metadata, download the NGS data / organism annotations, process them with preset pipelines for e.g. Ribo-seq, RNA-seq etc.
}
\author{
\strong{Maintainer}: Haakon Tjeldnes \email{hauken_heyken@hotmail.com} [data contributor]

Authors:
\itemize{
  \item Michal Swirski \email{michal.swirski@uw.edu.pl} [contributor]
}

Other contributors:
\itemize{
  \item Krzysztof Starecki \email{krzysztof.starecki@gmail.com} [contributor]
}

}
\keyword{internal}
