% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny_app_pipeline_monitor.R
\name{mNGSp_app}
\alias{mNGSp_app}
\title{Processing monitor app
Minimal Shiny dashboard for massiveNGSpipe runtime monitoring (full-width controls row)}
\usage{
mNGSp_app(
  config = pipeline_config(),
  pipelines_all = pipeline_init_all(config, only_complete_genomes = TRUE, gene_symbols =
    FALSE, show_status_per_exp = FALSE),
  options = list()
)
}
\description{
Processing monitor app
Minimal Shiny dashboard for massiveNGSpipe runtime monitoring (full-width controls row)
}
