% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipeline_utils.R
\name{name_of_function}
\alias{name_of_function}
\title{Get name of function}
\usage{
name_of_function(...)
}
\arguments{
\item{...}{functions}
}
\value{
character, name of function
}
\description{
Get name of function
}
