% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipeline_utils.R
\name{navigate}
\alias{navigate}
\title{Navigate to a Directory in the Files Pane}
\usage{
navigate(path)
}
\arguments{
\item{path}{The filesystem path to be shown.}
}
\value{
invisible(NULL)
}
\description{
Navigate to a directory in the Files pane. The contents of that directory
will be listed and shown in the Files pane.
}
