% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipeline_utils.R
\name{navigateExp}
\alias{navigateExp}
\title{Navigate to a Directory in the Files Pane}
\usage{
navigateExp(exp, rel_dir = "aligned")
}
\arguments{
\item{exp}{character, a ORFik experiment name of folder to go to, e.g. "GSE91068-saccharomyces_cerevisiae".}

\item{rel_dir}{character, relative directory, "aligned", "trim", "processed" or "raw"}
}
\value{
invisible(NULL)
}
\description{
Navigate to a directory in the Files pane.
The contents of that directory will be listed and shown in the Files pane.
}
