% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipeline_utils.R
\name{navigateRef}
\alias{navigateRef}
\title{Navigate to a reference Directory in the Files Pane}
\usage{
navigateRef(exp, rel_dir = "")
}
\arguments{
\item{exp}{ORFik experiment or character name of experiment,
a ORFik experiment name of folder to go to, e.g. "GSE91068-saccharomyces_cerevisiae".}

\item{rel_dir}{character, relative directory, default: ""}
}
\value{
invisible(NULL)
}
\description{
Navigate to a directory in the Files pane.
The contents of that directory will be listed and shown in the Files pane.
}
