% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipeline_preset_steps.R
\name{pipeline_collection_master}
\alias{pipeline_collection_master}
\title{Create the superset collection of all samples of all organisms
Also saves a file in pipeline dir called ./metadata_done_samples.csv.}
\usage{
pipeline_collection_master(
  config,
  pipelines = pipeline_init_all(config, gene_symbols = FALSE, only_complete_genomes =
    TRUE),
  done_experiments = step_is_done_pipelines(config, "pcounts", pipelines),
  done_organisms = unique(names(done_experiments)),
  path_suffix = "",
  BPPARAM = MulticoreParam(workers = bpparam()$workers, exportglobals = FALSE, log =
    FALSE, force.GC = FALSE, fallback = T)
)
}
\arguments{
\item{config}{list, output from pipeline_config(), the global config for your
NGS pipeline}

\item{pipelines}{list, output of pipeline_init_all}

\item{done_experiments}{named character vector, default:
step_is_done_pipelines(config, "pcounts", pipelines). Set of experiments to merge
per species.}

\item{done_organisms}{Scientific name (e.g. Homo sapiens), which organisms to run merging for, default
unique(names(done_experiments))}

\item{path_suffix}{character, default "". Add suffix to saved collection name, like a date to seperate version.}
}
\value{
invisible(NULL)
}
\description{
Create the superset collection of all samples of all organisms
Also saves a file in pipeline dir called ./metadata_done_samples.csv.
}
\examples{
pipeline_collection_org(config, done_organisms = "Homo sapiens", path_suffix = "10_02_2025")
}
