% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipeline_preset_steps_sub.R
\name{pipeline_create_experiment}
\alias{pipeline_create_experiment}
\title{Create ORFik experiment from config study and bam files}
\usage{
pipeline_create_experiment(pipeline, config)
}
\description{
Create ORFik experiment from config study and bam files
}
