% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipeline_preset_steps_sub.R
\name{pipeline_create_ofst}
\alias{pipeline_create_ofst}
\title{Create ORFik ofst files from bam files}
\usage{
pipeline_create_ofst(df_list, config)
}
\arguments{
\item{df_list}{a list of ORFik experiments}

\item{config}{the mNGSp config object from \link{pipeline_config}}
}
\description{
Create ORFik ofst files from bam files
}
