% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipeline_preset_steps_sub.R
\name{pipeline_download}
\alias{pipeline_download}
\title{Download all SRA files for all studies}
\usage{
pipeline_download(pipeline, config)
}
\arguments{
\item{pipeline}{a pipeline object}

\item{config}{the mNGSp config object from \link{pipeline_config}}
}
\value{
invisible(NULL)
}
\description{
Extract them into \verb{<accession>.fastq.gz} or \verb{<accession>_\{1,2\}.fastq.gz}
for SE/PE reads respectively.
}
