% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipeline_flags.R
\name{pipeline_flags}
\alias{pipeline_flags}
\title{Get all pipeline flags and set dirs}
\usage{
pipeline_flags(
  project_dir,
  flag_sub = c("start", "fetch", "trim", "collapsed", "aligned", "cleanbam", "exp",
    "ofst", "pshifted", "valid_pshift", "merged_lib", "covrle", "bigwig", "pcounts")
)
}
\arguments{
\item{project_dir}{pipeline directory}

\item{flag_sub}{character vector of names of flag steps to include}
}
\value{
a named character vector of directories in 'project_dir'
named as 'flag_sub' with names 'flag_sub'
}
\description{
Get all pipeline flags and set dirs
}
