% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipeline_steps_sub.R
\name{pipeline_init}
\alias{pipeline_init}
\title{Create a pipeline for the specified study.}
\usage{
pipeline_init(study, study_accession, config)
}
\arguments{
\item{study}{a data.table of output from ORFik::download.SRA.metadata}

\item{study_accession}{any accession accepted by
\code{ORFik::download.SRA.metadata}. It is needed to know
if you used GEO, PRJ, SRP etc as the search query.}

\item{config}{Configured directories for pipeline as a list}
}
\value{
a list of pipeline objects, one for each study,
subsetted by organism per study.
}
\description{
Create a pipeline for the specified study.
}
