% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipeline_steps.R
\name{pipeline_merge_org}
\alias{pipeline_merge_org}
\title{Merge all studies per organism}
\usage{
pipeline_merge_org(pipelines, config)
}
\arguments{
\item{pipelines}{list, output of pipeline_init_all}

\item{config}{list, output from pipeline_config(), the global config for your
NGS pipeline}
}
\value{
invisible(NULL)
}
\description{
Merge all studies per organism
}
