% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipeline_preset_steps.R
\name{pipeline_merge_org_modalities}
\alias{pipeline_merge_org_modalities}
\title{Gather all modalities per organism into exp}
\usage{
pipeline_merge_org_modalities(all_exp = list.experiments(validate = FALSE))
}
\arguments{
\item{all_exp}{data.table of all done experiment,
default: list.experiments(validate = FALSE)}
}
\value{
invisible(NULL)
}
\description{
Using libtype modalities like Ribo-seq, RNA-seq, disome-seq, create
a new experiment. Only applies to organisms with > 1 modality completed.
}
