% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipeline_metadata.R
\name{pipeline_validate_metadata}
\alias{pipeline_validate_metadata}
\title{Validate that metadata is ready to run}
\usage{
pipeline_validate_metadata(
  dt,
  config,
  output_file = config$complete_metadata,
  backup_file = config$backup_metadata,
  next_round_file = config$temp_metadata,
  libtypes = "RFP"
)
}
\arguments{
\item{dt}{a data.table of all metadata for all studies}

\item{config}{list, output from pipeline_config(), the global config for your
NGS pipeline}

\item{output_file}{a path to store final Ribo csv, default config$complete_metadata}

\item{backup_file}{a path to store backup csv, default: config$backup_metadata,
store all runs, even non Ribo-seq, such that it can be used if wanted.}
}
\description{
Validate that metadata is ready to run
}
