% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipeline_preset_steps_sub.R
\name{pipeline_validate_shifts}
\alias{pipeline_validate_shifts}
\title{Validate pshifting}
\usage{
pipeline_validate_shifts(df_list, config)
}
\arguments{
\item{df_list}{a list of ORFik experiments}

\item{config}{the mNGSp config object from \link{pipeline_config}}
}
\value{
invisible(NULL)
}
\description{
It will do these two things: Plot all TIS regions and
make a frame distribution table for all libs.
It then save a rds file called either 'warning.rds' or 'good.rds'.
Depending on the results was accepted or not:
\code{any(zero_frame$percent_length < 25)}
}
