% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipeline_logs.R
\name{progress_log}
\alias{progress_log}
\title{Get log for specific pipeline run}
\usage{
progress_log(
  config,
  task = 1,
  lines = 100,
  logdir = bplogdir(config$BPPARAM_MAIN)
)
}
\arguments{
\item{config}{a config object}

\item{task}{numeric, default 1. The task number wanted to check.}

\item{lines}{numeric, default 100. Number of lines to output (
starting from end of fail)}

\item{logdir}{character, default: bplogdir(config$BPPARAM),
path to log directory you want to check.}
}
\value{
NULL (text output to screen)
}
\description{
Get the n lasts line of a specific job for pipeline
}
