% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipeline_progress.R
\name{ps}
\alias{ps}
\title{Call View on get_running_processes}
\usage{
ps(pgid_subset = NULL)
}
\arguments{
\item{pgid_subset}{numeric, default NULL, else subset to process group}
}
\value{
either NULL or captured output
}
\description{
Nice to get an overview or kill a processes group by pgid
}
