% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny_app_reshifting.R
\name{psite_reshift_app}
\alias{psite_reshift_app}
\title{For each study that is done, validate / fix shifts}
\usage{
psite_reshift_app(
  config,
  exp_names = reshift_app_exp_names(config),
  options = list(launch.browser = TRUE)
)
}
\description{
For each study that is done, validate / fix shifts
}
