% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fastq_helpers.R
\name{read_fastq_adapter_output}
\alias{read_fastq_adapter_output}
\title{Read fastqc output (subset adapter)}
\usage{
read_fastq_adapter_output(qc_report_path)
}
\description{
Parse the report and return the adapter sequence with most hits
(or "disable" if none found). Threshold defined by fastqc and
"fail" means adapter was found.
}
