% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipeline_flags.R
\name{remove_flag_all_exp}
\alias{remove_flag_all_exp}
\title{Remove flags for all pipelines specified}
\usage{
remove_flag_all_exp(config, steps = names(config$flag), exps)
}
\arguments{
\item{config}{list, output from pipeline_config(), the global config for your
NGS pipeline}

\item{steps, }{which flags to set to FALSE (not done)
for given exp objects (i.e. a subset of pipelines)}

\item{exps}{name of experiments (i.e. subsets names of pipelines,
split by organisms.)}
}
\value{
invisible(NULL)
}
\description{
Set all flags specified as not done for all experiments in
for all pipeline objects specified. Remember this is exp subset of pipelines!
}
