% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fastq_helpers.R
\name{run_files_organizer}
\alias{run_files_organizer}
\title{Fastq/fastq file organizer}
\usage{
run_files_organizer(
  runs,
  source_dir,
  exclude = c("json", "html"),
  format = c(".fastq", ".fq", ".fa", ".fasta"),
  compressions = c("", ".gz")
)
}
\description{
Fastq/fastq file organizer
}
