% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipeline_run.R
\name{run_pipeline}
\alias{run_pipeline}
\title{Run massive_NGS_pipe}
\usage{
run_pipeline(pipelines, config, wait = 100)
}
\arguments{
\item{pipelines}{list, output of pipeline_init_all}

\item{config}{list, output from pipeline_config(), the global config for your
NGS pipeline}

\item{wait}{numeric, default 100 (in seconds). How long should each
partial pipeline wait if it done before it check for new results
ready to continue with.}
}
\value{
invisible(NULL)
}
\description{
Run massive_NGS_pipe
}
