% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collection_processing.R
\name{save_coverage_tiles}
\alias{save_coverage_tiles}
\title{Create megafst by coverage tiles}
\usage{
save_coverage_tiles(
  filepath,
  chromosomes,
  chromosome_lengths,
  output_dir,
  tile_size,
  colnames,
  verbose = TRUE,
  remake = FALSE,
  BPPARAM
)
}
\description{
Read + or - strand covRle over all libraries and process
}
