% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipeline_flags.R
\name{set_flag_all}
\alias{set_flag_all}
\title{Set flags for all pipelines given}
\usage{
set_flag_all(config, steps = names(config$flag), pipelines)
}
\arguments{
\item{config}{list, output from pipeline_config(), the global config for your
NGS pipeline}

\item{steps, }{which flags to set to TRUE for given pipeline objects}

\item{pipelines}{list, output of pipeline_init_all}
}
\value{
invisible(NULL)
}
\description{
Set all flags as done for all experiments in pipeline up to and included
the last step in steps
}
