% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipeline_flags.R
\name{set_flag_all_exp}
\alias{set_flag_all_exp}
\title{Set flags for all pipeline exp specified}
\usage{
set_flag_all_exp(config, steps = names(config$flag), exps)
}
\arguments{
\item{config}{list, output from pipeline_config(), the global config for your
NGS pipeline}

\item{steps, }{which flags to set to TRUE for given exp objects
(i.e. a subset of pipelines)}

\item{exps}{name of experiments (i.e. subsets names of pipelines,
split by organisms.)}
}
\value{
invisible(NULL)
}
\description{
Set all flags as done for all experiments in pipeline up to and included
the last step in steps. Remember this is exp subset of pipelines!
}
