% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipeline_flags.R
\name{step_is_done_pipelines}
\alias{step_is_done_pipelines}
\title{Check if step is done, pipeline input}
\usage{
step_is_done_pipelines(config, step, pipelines)
}
\arguments{
\item{config}{list, output from pipeline_config(), the global config for your
NGS pipeline}

\item{pipelines}{list, output of pipeline_init_all}
}
\value{
logical vector, TRUE if it is done.
}
\description{
Check if step is done, pipeline input
}
