% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/demultiplex.R
\name{umi_processing}
\alias{umi_processing}
\title{Process UMIs}
\usage{
umi_processing(reads, firstN, lastN, saveUMIs = FALSE)
}
\arguments{
\item{reads}{a list of DNAStringSet}

\item{firstN}{integer, number of nt for 5' UMI part}

\item{lastN}{integer, number of nt for 3' UMI part}

\item{saveUMIs}{logical, default FALSE.}
}
\value{
a list of DNAStringSet processed.
}
\description{
Process UMIs
}
