% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/demultiplex.R
\name{writeout.fastas}
\alias{writeout.fastas}
\title{Write a list of fasta files}
\usage{
writeout.fastas(reads, out_dir, append = FALSE, compress = FALSE)
}
\arguments{
\item{reads}{a named list of DNAStringSets, names are file prefixes,
will append .fasta or .fasta.gz depending on if compress is TRUE.}

\item{out_dir}{output directory}

\item{append}{logical, default FALSE. Append to file (FALSE) or Overwrite (TRUE)}

\item{compress}{logical, default FALSE. Compress to .gz or not.}
}
\value{
invisible(NULL), files saved to disc
}
\description{
Write a list of fasta files
}
