% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipeline_analysis.R
\name{analysis_pipeline_ribo_rna}
\alias{analysis_pipeline_ribo_rna}
\title{Analysis of Ribo-seq and RNA-seq}
\usage{
analysis_pipeline_ribo_rna(
  df.rfp = read.experiment("Eleonora-homo_sapiens"),
  df.rna = read.experiment("Eleonora-homo_sapiens_RNA-seq"),
  canonical_isoforms = canonical_isoforms(df.rfp)
)
}
\arguments{
\item{df.rfp}{an ORFik experiment}

\item{df.rna}{an ORFik experiment}

\item{canonical_isoforms}{a character vector with tx names,
default: canonical_isoforms(df.rfp)}
}
\value{
invisible(NULL), results saved to disc
}
\description{
Analysis of Ribo-seq and RNA-seq
}
