% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipeline_fail_reports.R
\name{bad_pshifting_report}
\alias{bad_pshifting_report}
\title{If pshifting fails, Check qc and shift plots}
\usage{
bad_pshifting_report(df)
}
\description{
If pshifting fails, Check qc and shift plots
}
