% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipeline_checklist.R
\name{checklist_path}
\alias{checklist_path}
\title{Path to the live checklist snapshot file for one project}
\usage{
checklist_path(config)
}
\arguments{
\item{config}{the mNGSp config object}
}
\value{
character, file.path(config$project, "log_pipeline", "checklist.txt")
}
\description{
Path to the live checklist snapshot file for one project
}
