% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipeline_utils.R
\name{copy_experiments_to}
\alias{copy_experiments_to}
\title{Copy done experiment csvs to new directory}
\usage{
copy_experiments_to(
  csv_names,
  old_exp_dir = ORFik::config()["exp"],
  new_exp_dir = sub("_local$|_local/$", "", old_exp_dir),
  docker_conversion = "'s/livemount\\\\///g'"
)
}
\description{
Copy done experiment csvs to new directory
}
\examples{
csvs <- list.files(ORFik::config()["exp"], pattern = "modalities\\\\.csv$")
copy_experiments_to(csvs)
}
