% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipeline_metadata.R
\name{curate_metadata}
\alias{curate_metadata}
\title{Curate metadata for massive_NGS_pipe}
\usage{
curate_metadata(
  accessions,
  config,
  organisms = "all",
  google_url = config$google_url,
  complete_metadata = config$complete_metadata,
  LibraryLayouts = c("SINGLE", "PAIRED"),
  LibraryStrategy = c("RNA-Seq", "miRNA-Seq", "OTHER"),
  libtypes = "RFP",
  Platforms = "ILLUMINA",
  step_mode = FALSE,
  open_google_sheet = interactive(),
  fix_loop = TRUE,
  only_curated = FALSE
)
}
\arguments{
\item{accessions}{character vector, all candidate accession numbers, allowed types:\cr
\itemize{
\item Biorpoject ID (Only numbers)\cr
\item Bioproject accession (PRJ)\cr
\item SRA study (SRA)\cr
\item ENA study (ERA)\cr
\item GEO study (GSE)\cr
\item NULL (not "NULL", but NULL), don't add new accession, only look if you
current google sheet or local csv is ready.
}}

\item{config}{path, default \code{ORFik::config()}, where will
fastq, bam, references and ORFik experiments go}

\item{organisms}{character vector, default "all" (Use all organisms found).
Else binomial latin name with capital letter for genus: "Homo sapiens" etc.}

\item{google_url}{url or sheet object for google sheet to use. Set to NULL to
not use google sheet.}

\item{complete_metadata}{path, default file.path(project_dir, "RFP_FINAL_LIST.csv").
The list of final candidates that are checked and have unique rows per bioproject}

\item{LibraryLayouts}{character vector, default c("SINGLE", "PAIRED"),
either or both of: c("SINGLE", "PAIRED")}

\item{LibraryStrategy}{character vector, default:
c("RNA-Seq", "miRNA-Seq", "OTHER"). Which library Strategies to keep.
Default is the total set of normal RNA sequencing. (RNA-seq, miRNA-seq,
Ribo-seq, CHIP-seq, CAGE etc.) Change if you only want a subset, or some
rare type.}

\item{libtypes}{character vector, default "RFP". Which NGS library types to
keep from all LIBRARYLAYOUT values. These names are made by mNGSp
automatically. If you made "disome" as a LIBRARYLAYOUT category, then
c("RFP", "disome") will keep both, but exclude any other.}

\item{Platforms}{character vector, default: "ILLUMINA". The sequencer technologies allowed.}

\item{step_mode}{logical, default FALSE. If TRUE, actives browser() and lets you go through
each step in debug mode for full control.}

\item{fix_loop}{logical, default TRUE. If TRUE, will create a loop in
the end that runs until your metadata is valid. If you
only want to updata csv / google docs with new data and not try to complete
now, set this to FALSE.}

\item{only_curated}{logical FALSE, if TRUE. Only validate (or fail) based on
inserted accessions (not the full list). Ignored if accessions is NULL.}
}
\value{
logical, TRUE if you had unique rows per accession per organism
}
\description{
Runs an interactive while loop until your metadata is curated and have this property:
Per accession, your study must have a unique set of rows per organism.
Such that ORFik experiment can understand what is what:
(what is Wild type vs experimental), which are replicates etc.
The uniqueness is defined from column LIBRARYTYPE to column INHIBITOR
}
