% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/google_API.R
\name{discord_connection_default_cached}
\alias{discord_connection_default_cached}
\title{Get discord default webhook}
\usage{
discord_connection_default_cached(
  channel_id = 1,
  connection_file = "~/livemount/.cache/discordr/r_to_discord_config",
  verbose = FALSE
)
}
\arguments{
\item{channel_id}{numeric, id of channel to use.}

\item{connection_file}{character, file path to cached webhooks}

\item{verbose}{logical, FALSE. If TRUE, will display discordr info.}
}
\value{
list of size 4, a webhook object
}
\description{
Get discord default webhook
}
