% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipeline_utils.R
\name{docker_copy_done_experiments}
\alias{docker_copy_done_experiments}
\title{For docker usage copy}
\usage{
docker_copy_done_experiments(
  config,
  pipelines = pipeline_init_all(config),
  new_exp_dir = "~/livemount/Bio_data/ORFik_experiments/",
  docker_conversion = "'s/livemount\\\\///g'",
  individual_studies = TRUE,
  merged_by_organism = TRUE,
  all_by_organism = TRUE,
  all_modalities_by_org = TRUE
)
}
\arguments{
\item{config}{a config object}

\item{pipelines}{pipeline objects, default pipeline_init_all(config)}

\item{new_exp_dir}{character, default "~/livemount/Bio_data/ORFik_experiments/"}

\item{docker_conversion}{the sed conversion string: 's/livemount\///g'}

\item{individual_studies}{logical, default TRUE. Should raw ORFik experiments
be exported per study.}

\item{merged_by_organism}{logical, default TRUE.
Create a new study, merge the superset of all per organism into single file.
This requires that you have run: pipeline_merge_org}

\item{all_by_organism}{logical, default TRUE.
Create a new study, the superset of all studies of all per organism.
This requires that you have run: pipeline_merge_org}

\item{all_modalities_by_org}{logical, default TRUE.
Should the all modalities tracks be exported per organism.}
}
\value{
logical, TRUE if sucessful for all.
}
\description{
If you have download docker and release docker.
This is useful to copy files from download to release
}
