% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/google_API.R
\name{download_fastq_google_drive}
\alias{download_fastq_google_drive}
\title{Get all fastq files from google drive}
\usage{
download_fastq_google_drive(
  drive_url,
  output_dir,
  files = google_drive_list_files(drive_url, email, format_pattern, recursive),
  format_pattern = "fastq\\\\.gz$",
  email = gargle_mail_safe(),
  overwrite = FALSE,
  recursive = TRUE
)
}
\arguments{
\item{drive_url}{character, url to drive folder (not individual files), example:
"https://drive.google.com/drive/folders/1N_Tyig3-IxMIe8Im24wBSNiS5vSohi91"}

\item{output_dir}{character, output directory, will be create if not existing}

\item{email}{character, email of user, default: \email{hakontjeldnes@gmail.com}}

\item{overwrite}{logical, default FALSE. If FALSE, delete existing files,
if TRUE dont download existing files.}

\item{recursive}{logical, default TRUE. Search sub folders for format files.}

\item{pattern}{character pattern of files, default "fastq\\.gz$"}
}
\value{
invisible(NULL)
}
\description{
Get all fastq files from google drive
}
