% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fastq_helpers.R
\name{fastqc_adapters_info}
\alias{fastqc_adapters_info}
\title{Helper function to find the main adapter in a .fastq.gz file}
\usage{
fastqc_adapters_info(file)
}
\arguments{
\item{file}{full path to file to check with qc}
}
\value{
adapter candidates
}
\description{
Helper function to find the main adapter in a .fastq.gz file
}
