% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipeline_init_helpers.R
\name{get_annotation}
\alias{get_annotation}
\title{Get all required annotation for an organism}
\usage{
get_annotation(
  organism,
  output.dir = reference_folder_name(organism, TRUE),
  genome = TRUE,
  GTF = TRUE,
  db = "ensembl",
  gene_symbols = TRUE
)
}
\arguments{
\item{organism}{scientific name of organism, Homo sapiens,
Danio rerio, Mus musculus, etc. See \code{biomartr:::get.ensembl.info()}
for full list of supported organisms.}

\item{output.dir}{path, default: reference_folder_name(organism, TRUE)}

\item{genome}{logical, default: TRUE. Download the genome for the organism
given in `organism`. If `FALSE`, ORFik checks whether the file is already
available locally. To use a custom genome from disk, set
\code{genome = FALSE} and then assign:\cr
\code{annotation <- getGenomeAndAnnotation(genome = FALSE)}\cr
\code{annotation["genome"] = "path/to/genome.fasta"}.\cr
For Ensembl, the primary assembly is downloaded.}

\item{GTF}{logical, default: TRUE. Download the annotation file for the
organism given in `organism`. If `FALSE`, ORFik checks whether the file is
already available locally. To use a custom annotation from disk, set
`GTF = FALSE` and then assign:\cr
annotation <- getGenomeAndAnnotation(gtf = FALSE)\cr
annotation["gtf"] = "path/to/gtf.gtf".\cr
If `db` is not `"ensembl"`, the downloaded annotation will be GFF instead
of GTF.}

\item{db}{database to use for the genome and annotation. The recommended
choice is `"ensembl"`. When using Ensembl, remember to set
\code{assembly_type = "primary_assembly"} unless you want haplotypes
included, which can make the files much larger. Alternatives are `"refseq"`
for reference assemblies and `"genbank"` for all available assemblies.}

\item{gene_symbols}{logical default FALSE. If TRUE, will download
and store all gene symbols for all transcripts (coding and noncoding)-
In a file called: "gene_symbol_tx_table.fst" in same folder as txdb.
hgcn for human, mouse symbols for mouse and rat, more to be added.}
}
\value{
the annotation paths as vector
}
\description{
Get: Fasta genome, transcriptome annotation (gff/gtf),
contaminants, ORFik speedup objects, gene symbols,
uniprot ids (to be added).
For local annotation, set genome and GTF arguments to full
path of respective objects.
}
