% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipeline_init_helpers.R
\name{get_annotation}
\alias{get_annotation}
\title{Get all required annotation for an organism}
\usage{
get_annotation(
  organism,
  output.dir = file.path(ORFik::config()["ref"], organism),
  genome = TRUE,
  GTF = TRUE,
  db = "ensembl",
  gene_symbols = TRUE
)
}
\arguments{
\item{organism}{scientific name of organism, Homo sapiens,
Danio rerio, Mus musculus, etc. See \code{biomartr:::get.ensembl.info()}
for full list of supported organisms.}

\item{output.dir}{path, file.path(ORFik::config()\link{"ref"}, organism)}

\item{genome}{logical, default: TRUE, download genome of organism
specified in "organism" argument. If FALSE, check if the downloaded
file already exist. If you want to use a custom gtf from you hard drive,
set \code{GTF = FALSE},
and assign: \cr \code{annotation <- getGenomeAndAnnotation(genome = FALSE)}\cr
\code{annotation["genome"] = "path/to/genome.fasta"}.\cr
Will download the primary assembly from Ensembl.}

\item{GTF}{logical, default: TRUE, download gtf of organism specified
in "organism" argument. If FALSE, check if the downloaded
file already exist. If you want to use a custom gtf from you hard drive,
set GTF = FALSE,
and assign: \cr annotation <- getGenomeAndAnnotation(gtf = FALSE)\cr
annotation["gtf"] = "path/to/gtf.gtf".\cr
If db is not "ensembl", you will instead get a gff file.}

\item{db}{database to use for genome and GTF,
default adviced: "ensembl" (remember to set assembly_type to "primary_assembly",
else it will contain haplotypes, very large file!).
Alternatives: "refseq" (primary assembly) and "genbank" (mix)}

\item{gene_symbols}{logical default FALSE. If TRUE, will download
and store all gene symbols for all transcripts (coding and noncoding)-
In a file called: "gene_symbol_tx_table.fst" in same folder as txdb.
hgcn for human, mouse symbols for mouse and rat, more to be added.}
}
\value{
the annotation paths as vector
}
\description{
Get: Fasta genome, transcriptome annotation (gff/gtf),
contaminants, ORFik speedup objects, gene symbols,
uniprot ids (to be added).
For local annotation, set genome and GTF arguments to full
path of respective objects.
}
