% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipeline_progress.R
\name{get_running_processes}
\alias{get_running_processes}
\title{Get running processes}
\usage{
get_running_processes(
  fields = "pid,pgid,user,pcpu,pmem,nice,priority,time,command",
  as.data.table = FALSE
)
}
\arguments{
\item{fields}{"pid,pgid,user,pcpu,pmem,nice,priority,time,command"}

\item{as.data.table}{logical, default FALSE. If FALSE, capture as character vector, else
convert to data.table.}
}
\value{
either NULL or captured output
}
\description{
Nice to kill a processes group by pgid
}
