% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipeline_progress.R
\name{get_running_processes}
\alias{get_running_processes}
\title{Get running processes}
\usage{
get_running_processes(
  pgid_subset = NULL,
  as.data.table = TRUE,
  fields = "pid,pgid,user,pcpu,pmem,nice,priority,time,command"
)
}
\arguments{
\item{pgid_subset}{numeric, default NULL, else subset to process group}

\item{as.data.table}{logical, default T If FALSE, capture as character vector, else
convert to data.table.}

\item{fields}{"pid,pgid,user,pcpu,pmem,nice,priority,time,command"}
}
\value{
either NULL or captured output
}
\description{
Nice to get an overview or kill a processes group by pgid
}
