% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipeline_logs.R
\name{last_update}
\alias{last_update}
\title{Get latest time of pipeline update flag}
\usage{
last_update(config, flagdir = dirname(config$flag[1]))
}
\arguments{
\item{flagdir}{character, default: dirname(config$flag\link{1}),
path to flag directory}

\item{a}{config object}
}
\value{
character, a date string
}
\description{
Get latest time of pipeline update flag
}
