% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipeline_metadata.R
\name{local_study_csv}
\alias{local_study_csv}
\title{Create template study csv info}
\usage{
local_study_csv(
  dir,
  name,
  files = list.files(dir, pattern = "\\\\.fast.*", full.names = TRUE),
  organism,
  sample_title,
  paired = FALSE,
  Model = "Illumina Genome Analyzer",
  Platform = "ILLUMINA",
  LibraryName = "",
  LibraryStrategy = "RNA-Seq",
  LibrarySource = "TRANSCRIPTOMIC",
  SampleName = c(""),
  sample_source = c(""),
  MONTH = substr(Sys.time(), 6, 7),
  YEAR = substr(Sys.time(), 1, 4),
  AUTHOR = Sys.info()["user"]
)
}
\arguments{
\item{dir}{a directory with fast files
(fasta or fastq, uncompressed/compressed)}

\item{name}{a basis for directory name. The relative path without
the -organism part.}

\item{files}{character vector of full paths,
default: \code{list.files(dir, pattern = "\\.fast.*", full.names = TRUE)}}

\item{organism}{Scientific latin name of organism, give single if all
are equal, or specify for each file if multiple. Example: "Homo sapiens"}

\item{sample_title}{Define a valid sample name per sample, like,
'WT_ribo_cyclohexamide_rep1'}

\item{paired}{character, either "SINGLE" or "PAIRED"}

\item{Model}{Sequencing machine name, default: "Illumina Genome Analyzer",
give single if all are equal, or specify for each file if multiple.}

\item{Platform}{Sequencing machine platform, default: "ILLUMINA",
give single if all are equal, or specify for each file if multiple.}
}
\value{
a data.table with required columns:\cr
study_accession: the basename of folder for experiment\cr
RUN: The sample name (For paired end, still 1 row, with basename of sample)\cr
...
}
\description{
Create template study csv info
}
