% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipeline_collapse.R
\name{pipe_collapse}
\alias{pipe_collapse}
\title{Collapse trimmed single-end reads and move them into "trim/SINGLE"
subdirectory. PE reads are moved into "trim/PAIRED" without modification.}
\usage{
pipe_collapse(pipeline, config)
}
\arguments{
\item{pipeline}{a pipeline object}
}
\description{
Collapse trimmed single-end reads and move them into "trim/SINGLE"
subdirectory. PE reads are moved into "trim/PAIRED" without modification.
}
