% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipeline_preset_steps_sub.R
\name{pipeline_align_contaminants}
\alias{pipeline_align_contaminants}
\title{Remove contaminants and align the reads to genome. Single and paired end
reads are handled separately, so the resulting logs are renamed with a
"_SINGLE" and/or "_PAIRED" suffix.}
\usage{
pipeline_align_contaminants(pipeline, config)
}
\arguments{
\item{pipeline}{a pipeline object}

\item{config}{the mNGSp config object from \link{pipeline_config}}
}
\description{
Remove contaminants and align the reads to genome. Single and paired end
reads are handled separately, so the resulting logs are renamed with a
"_SINGLE" and/or "_PAIRED" suffix.
}
