% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipeline_steps_sub.R
\name{pipeline_cleanup}
\alias{pipeline_cleanup}
\title{Remove all files apart from logs and final aligned BAMs.
Rename the BAMs into <run_accession>.bam format.}
\usage{
pipeline_cleanup(pipeline, config)
}
\arguments{
\item{pipeline}{a pipeline object}
}
\description{
Remove all files apart from logs and final aligned BAMs.
Rename the BAMs into <run_accession>.bam format.
}
