% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipeline_collapse.R
\name{pipeline_collapse}
\alias{pipeline_collapse}
\title{Collapse fastq/fasta
Collapse trimmed single-end fasta/fastq reads and move them into "trim/SINGLE"
subdirectory. PE reads are moved into "trim/PAIRED" without modification.}
\usage{
pipeline_collapse(pipeline, config)
}
\arguments{
\item{pipeline}{a pipeline object, subset of init_pipelines output}

\item{config}{a pipeline_config object}
}
\description{
Collapse fastq/fasta
Collapse trimmed single-end fasta/fastq reads and move them into "trim/SINGLE"
subdirectory. PE reads are moved into "trim/PAIRED" without modification.
}
