% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipeline_preset_steps_sub.R
\name{pipeline_collapse}
\alias{pipeline_collapse}
\title{Collapse trimmed single-end reads and move them into "trim/SINGLE"
subdirectory. PE reads are moved into "trim/PAIRED" without modification.}
\usage{
pipeline_collapse(pipeline, config)
}
\arguments{
\item{pipeline}{a pipeline object}
}
\description{
Collapse trimmed single-end reads and move them into "trim/SINGLE"
subdirectory. PE reads are moved into "trim/PAIRED" without modification.
}
