% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipeline_preset_steps.R
\name{pipeline_collection_org}
\alias{pipeline_collection_org}
\title{Create the superset collection of all samples per organism}
\usage{
pipeline_collection_org(
  config,
  pipelines = pipeline_init_all(config, gene_symbols = FALSE, only_complete_genomes =
    TRUE),
  done_experiments = step_is_done_pipelines(config, "pcounts", pipelines),
  done_organisms = unique(names(done_experiments)),
  path_suffix = "",
  BPPARAM = MulticoreParam(workers = bpparam()$workers, exportglobals = FALSE, log =
    FALSE, force.GC = FALSE, fallback = T)
)
}
\arguments{
\item{config}{list, output from pipeline_config(), the global config for your
NGS pipeline}

\item{pipelines}{list, output of pipeline_init_all}

\item{done_experiments}{named character vector, default:
step_is_done_pipelines(config, "merged_lib", pipelines). Set of experiments to merge
per species.}

\item{done_organisms}{Scientific name (e.g. Homo sapiens), which organisms to run merging for, default
unique(names(done_experiments))}

\item{path_suffix}{character, default "". Add suffix to saved collection name, like a date to seperate version.}
}
\value{
invisible(NULL)
}
\description{
Create the superset collection of all samples per organism
}
\examples{
config <- pipeline_config()
pipelines <- pipeline_init_all(config, gene_symbols = FALSE, only_complete_genomes = TRUE)
dirs <- lapply(pipelines, function(p) {
dirs_pipe <- massiveNGSpipe:::list_dirs_of_pipeline(pipeline = p)
names(dirs_pipe) <- tolower(names(dirs_pipe))
lapply(dirs_pipe, function(exp) {
  all(dir.exists(exp[-which(names(exp) \%in\% c("fastq", "trim", "exp"))]))
})
})
exists <- unlist(dirs, use.names = TRUE)
names(exists) <- sub("\\\\.", "-", names(exists))
names(exists) <- sub(" ", "_", names(exists))
sum(exists / length(exists))
all_exp_names <- unlist(massiveNGSpipe:::get_experiment_names(pipelines), use.names = FALSE)
stopifnot(all(names(exists) \%in\% all_exp_names))
done_experiments <- names(exists)[exists]
length(done_experiments)
names(done_experiments) <- stringr::str_to_sentence(sub("_", " ", sub(".*-", "", done_experiments)))
pipeline_collection_org(config, pipelines, done_organisms = "Homo sapiens", done_experiments, path_suffix = "_13_06_2025")
#pipeline_collection_org(config, done_organisms = "Homo sapiens", path_suffix = "_10_02_2025")
}
