% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipeline_config.R
\name{pipeline_config}
\alias{pipeline_config}
\title{Initial path config}
\usage{
pipeline_config(
  project_dir = file.path(dirname(config)[1], "NGS_pipeline"),
  config = ORFik::config()
)
}
\arguments{
\item{project_dir}{where will specific pipeline outputs be put}

\item{config}{path, default \code{ORFik::config()}, where will
fastq, bam, references and ORFik experiments go}
}
\value{
a list with a defined config
}
\description{
Set up all paths and flag directories
}
