% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipeline_flags.R
\name{pipeline_flags}
\alias{pipeline_flags}
\title{Get all pipeline flags and set dirs}
\usage{
pipeline_flags(
  project_dir,
  mode = c("online", "local")[1],
  preset,
  contam = FALSE,
  create_dirs = TRUE,
  flag_names = libtype_flags(preset, mode, contam)
)
}
\arguments{
\item{project_dir}{pipeline directory}

\item{mode}{= \code{c("online", "local")[1]}. "online" will assume project IDs for
online repository (SRA, ENA, PRJ etc). Local means local folders as accessions.}

\item{preset}{character, default "Ribo-seq".
Alternatives: c("Ribo-seq", "RNA-seq", "disome", "empty")}

\item{contam}{logical, FALSE, do contamint removal, using a seperat made
STAR index during genome preparation step.}

\item{create_dirs}{logical, default TRUE. For test purpose you can turn of
dir creation if needed.}

\item{flag_names}{named character vector of names of flag steps to include.
Names are the grouping inside functions and values are individual steps.}
}
\value{
a named character vector of directories in 'project_dir'
named as 'flag_sub' with names 'flag_sub'
}
\description{
Get all pipeline flags and set dirs
}
