% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipeline_flags.R
\name{pipeline_flags}
\alias{pipeline_flags}
\title{Get all pipeline flags and set dirs}
\usage{
pipeline_flags(
  project_dir,
  mode = c("online", "local")[1],
  preset,
  flag_names = libtype_flags(preset, mode),
  create_dirs = TRUE
)
}
\arguments{
\item{project_dir}{pipeline directory}

\item{mode}{= c("online", "local")\link{1}. "online" will assume project IDs for
online repository (SRA, ENA, PRJ etc). Local means local folders as accessions.}

\item{flag_names}{character vector of names of flag steps to include}

\item{create_dirs}{logical, default TRUE. For test purpose you can turn of
dir creation if needed.}
}
\value{
a named character vector of directories in 'project_dir'
named as 'flag_sub' with names 'flag_sub'
}
\description{
Get all pipeline flags and set dirs
}
