% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipeline_flags.R
\name{pipeline_flags}
\alias{pipeline_flags}
\title{Get all pipeline flags and set dirs}
\usage{
pipeline_flags(
  project_dir,
  mode = c("online", "local")[1],
  flag_sub = c("start", "fetch", "trim", "collapsed", "aligned", "cleanbam", "exp",
    "ofst", "pshifted", "valid_pshift", "merged_lib", "covrle", "bigwig", "pcounts"),
  flags_to_use = if (mode == "online") {
     seq(length(flag_sub))
 } else
    seq(length(flag_sub))[-c(1, 2)]
)
}
\arguments{
\item{project_dir}{pipeline directory}

\item{mode}{= c("online", "local")\link{1}. "online" will assume project IDs for
online repository (SRA, ENA, PRJ etc). Local means local folders as accessions.}

\item{flag_sub}{character vector of names of flag steps to include}

\item{flags_to_use}{numeric, vector of indices of 'flag_sub' to use.
c(1,2), would mean only download.}
}
\value{
a named character vector of directories in 'project_dir'
named as 'flag_sub' with names 'flag_sub'
}
\description{
Get all pipeline flags and set dirs
}
