% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipeline_steps.R
\name{pipeline_init_all}
\alias{pipeline_init_all}
\title{Init all studies to pipeline objects}
\usage{
pipeline_init_all(
  config,
  complete_metadata = config$complete_metadata,
  simple_progress_report = TRUE
)
}
\arguments{
\item{config}{list, output from pipeline_config(), the global config for your
NGS pipeline}

\item{complete_metadata}{path, default file.path(project_dir, "RFP_FINAL_LIST.csv").
The list of final candidates that are checked and have unique rows per bioproject}

\item{simple_progress_report}{logical, default TRUE. Display current progress.}
}
\value{
a list of pipelines
}
\description{
Init all studies to pipeline objects
}
