% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipeline_preset_steps.R
\name{pipeline_merge_org}
\alias{pipeline_merge_org}
\title{Merge all studies per organism}
\usage{
pipeline_merge_org(config, pipelines = pipeline_init_all(config))
}
\arguments{
\item{config}{list, output from pipeline_config(), the global config for your
NGS pipeline}

\item{pipelines}{list, output of pipeline_init_all}
}
\value{
invisible(NULL)
}
\description{
Merge all studies per organism
}
