% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipeline_preset_steps.R
\name{pipeline_merge_org_modalities}
\alias{pipeline_merge_org_modalities}
\title{Gather all modalities per organism into exp}
\usage{
pipeline_merge_org_modalities(
  all_exp = list.experiments(validate = FALSE),
  out_dir_base = ORFik::config()["bam"]
)
}
\arguments{
\item{all_exp}{data.table of all done experiment,
default: list.experiments(validate = FALSE)}

\item{out_dir_base}{= ORFik::config()\link{"bam"}, will make directory internally
called name of exp (e.g. all_merged-Homo_sapiens_modalities)}
}
\value{
invisible(NULL)
}
\description{
Using libtype modalities like Ribo-seq, RNA-seq, disome-seq, create
a new experiment. Only applies to organisms with > 1 modality completed.
}
