% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipeline_metadata.R
\name{pipeline_metadata}
\alias{pipeline_metadata}
\title{Download all metadata for all accessions}
\usage{
pipeline_metadata(accessions, config, force = FALSE)
}
\arguments{
\item{accessions}{character vector, all candidate accession numbers, allowed types:\cr
\itemize{
\item Biorpoject ID (Only numbers)\cr
\item Bioproject accession (PRJ)\cr
\item SRA study (SRA)\cr
\item ENA study (ERA)\cr
\item GEO study (GSE)\cr
}}

\item{config}{list, output from pipeline_config(), the global config for your
NGS pipeline}

\item{force}{logical, default FALSE. Force redownload of all, if something failed}
}
\value{
a data.table of all metadata for all accessions
}
\description{
Download all metadata for all accessions
}
