% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipeline_steps_sub.R
\name{pipeline_trim}
\alias{pipeline_trim}
\title{Trim all .fastq.gz files in a given pipeline. Paired end reads
("SRR<...>_1.fastq.gz", "SRR<...>_2.fastq.gz") are trimmed separately
from each other to allow for differing adapters.}
\usage{
pipeline_trim(pipeline, config)
}
\arguments{
\item{pipeline}{a pipeline object}
}
\description{
Trim all .fastq.gz files in a given pipeline. Paired end reads
("SRR<...>_1.fastq.gz", "SRR<...>_2.fastq.gz") are trimmed separately
from each other to allow for differing adapters.
}
