% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipeline_metadata.R
\name{pipeline_validate_metadata}
\alias{pipeline_validate_metadata}
\title{Validate that metadata is ready to run}
\usage{
pipeline_validate_metadata(
  dt,
  config,
  output_file = file.path(config$project, "RFP_FINAL_LIST.csv"),
  backup_file = file.path(config$project, "ALL_BACKUP_LIST.csv")
)
}
\arguments{
\item{dt}{a data.table of all metadata for all studies}

\item{config}{list, output from pipeline_config(), the global config for your
NGS pipeline}

\item{output_file}{a path to store final Ribo csv}

\item{backup_file}{a path to store backup csv, store all runs, even non Ribo-seq, such that it can be used if wanted.}
}
\description{
Validate that metadata is ready to run
}
