% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipeline_validate.R
\name{progress_report}
\alias{progress_report}
\title{Report summary of current progress}
\usage{
progress_report(
  pipelines,
  config,
  show_stats = TRUE,
  show_done = TRUE,
  status_plot = FALSE
)
}
\arguments{
\item{pipelines}{list, output of pipeline_init_all}

\item{config}{list, output from pipeline_config(), the global config for your
NGS pipeline}

\item{show_stats}{logical, default TRUE, output trim/alignment stats
\itemize{
\item plots, set to FALSE if you only want progress report.
}}

\item{show_done}{logical, default TRUE. If FALSE, display only status
of projects that are not done. Stats will still show for all.}
}
\value{
invisible(NULL)
}
\description{
How many studies have finished etc
}
