% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipeline_validate.R
\name{progress_report}
\alias{progress_report}
\title{Report summary of current progress}
\usage{
progress_report(pipelines, config, show_stats = TRUE)
}
\arguments{
\item{pipelines}{list, output of pipeline_init_all}

\item{config}{list, output from pipeline_config(), the global config for your
NGS pipeline}

\item{show_stats}{logical, default TRUE, output trim/alignment stats
\itemize{
\item plots, set to FALSE if you only want progress report.
}}
}
\value{
invisible(NULL)
}
\description{
How many studies have finished etc
}
