% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipeline_progress.R
\name{status_per_study}
\alias{status_per_study}
\title{Status per study}
\usage{
status_per_study(pipelines, config, show_status_per_exp, show_done, show_stats)
}
\arguments{
\item{pipelines}{list, output of pipeline_init_all}

\item{config}{list, output from pipeline_config(), the global config for your
NGS pipeline}

\item{show_status_per_exp}{logical, default TRUE.
Show a message per experiment with progression status.}

\item{show_done}{logical, default TRUE. If FALSE, display only status
of projects that are not done. Stats will still show for all.}

\item{show_stats}{logical, default FALSE, if TRUE output trim/alignment stats
\itemize{
\item plots, FALSE if you only want progress report and save some time.
}}
}
\value{
a list structured as: list(done, progress_index, projects,
alignment.stats.all, trimmed.out.all, n_bioprojects,
report_dir = config$project, message, steps, all_organism)
}
\description{
Status per study
}
