% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipeline_progress.R
\name{status_plot}
\alias{status_plot}
\title{Status plot of pipeline}
\usage{
status_plot(steps, progress_index, projects, n_bioprojects, done)
}
\description{
Given all studies, report how far they have come as a heatmap (y-axis
is steps, x-axis is studies)
}
