% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/google_API.R
\name{sync_sheet_safe}
\alias{sync_sheet_safe}
\title{Sync your local temp to google sheet}
\usage{
sync_sheet_safe(config, validate_to_complete_local = FALSE)
}
\arguments{
\item{config}{a mNGSp config with defined google_url and existing
temp_metadata file}
}
\value{
invisible(NULL)
}
\description{
Sync your local temp to google sheet
}
